Make Clippy happy again by removing supposedly unnecessary casts.

+++ pyo3-ffi/src/cpython/listobject.rs
@@ -24,14 +24,14 @@ pub struct PyListObject {
 #[inline]
 #[cfg(not(PyPy))]
 pub unsafe fn PyList_GET_ITEM(op: *mut PyObject, i: Py_ssize_t) -> *mut PyObject {
-    *(*(op as *mut PyListObject)).ob_item.offset(i as isize)
+    *(*(op as *mut PyListObject)).ob_item.offset(i)
 }
 
 /// Macro, *only* to be used to fill in brand new lists
 #[inline]
 #[cfg(not(PyPy))]
 pub unsafe fn PyList_SET_ITEM(op: *mut PyObject, i: Py_ssize_t, v: *mut PyObject) {
-    *(*(op as *mut PyListObject)).ob_item.offset(i as isize) = v;
+    *(*(op as *mut PyListObject)).ob_item.offset(i) = v;
 }
 
 #[inline]
